+++ CHANGELOG.md
@@ -375,3 +375,4 @@
 ### v2.0 12/15/19
 * changed airmon-ng stop/start exe time in network_crack to specific options
 * fixed git issues
+* removed sleep calls in airgeddon and add hashcat-utils make to lib_install in airgeddon/